Rename dotation directories to dotations
